linux/perf: Avoid returning ring-buffer memory
Use per-Event storage for records returned by ReadRecord/ReadRawRecord
instead of returning a slice that uses ring-buffer memory as storage.

Otherwise a record can be overwritten by the kernel while still in use.

This patch also includes fail-safe decoding of records and the ability
of resetting the ring-buffer shall it become corrupted. But the main fix
makes this probably unnecessary.